Database/Garrison : Fix mistake

+++ sql/updates/characters/2016_05_27_character_garrison_weekly_tavern_data.sql
@@ -1,5 +1,5 @@
 create table `character_garrison_weekly_tavern_data` (
 	`CharacterGUID` int (10),
-	`FollowerID` smallint (4),
+	`FollowerID` INT(10),
 	`Abilities` blob 
 );
